pyproject.toml
[project]
name = "reference-importer"
version = "2.0.0"
requires-python = ">= 3.9"

authors = [
  {name = "Jaime Florian", email = "jaimeflorian27@gmail.com"}
]

dependencies = [
  "ffmpeg-python==2",
  "Qt5.py==0.1.0",
]

[project.optional-dependencies]
dev = ["ruff<0.4.0"]

target-version = ["py39"]
[tool.ruff.lint]
select=["ALL"]
ignore = ["ANN101", "ANN102", "S603", "FBT003", "FBT001"]

[tool.ruff.lint.pydocstyle]
convention = "google"  

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]
"test_*.py" = ["S101"]
